requirements.txt
ansible==1.3
SQLAlchemy==0.8.1
Flask-SQLAlchemy==0.16
